Fix for laser evolution skipping some time steps
Before it could happen that in cases where the runge-kutta time step
was larger or similar than `dz_fields`, the wakefield would not be
calculated at all steps. This would then lead to `laser.evolve` being
called less times than needed, and this to a laser which has
evolved less than it should.