Add laser envelope model to fluid wakefield
Lengthy commit which includes the following changes:
* Replace old laser implementation with the new envelope model.
* Include `dz_fields` attribute to have the same functionality as in
`Quasistatic2DWakefield`
* Change the internal logic which determines when a new wakefield
calculation should be performed, when should the fields be
interpolated to particles, and how the laser envelope is evolved.
This mirrors the implemetation in `Quasistatic2DWakefield`. At some
point this could be generalized to all Wakefield classes.
* Fix inaccuracy in `dz` calculation (it was assuming cell centered,
but should be node centered).
* `z_arr` is now initialized with the field positions and does not have
to be filled in during the main loop.
* Added calculation of `rho` and `chi`.
* (Older change) Adapt the field gathering to the new implementation
of `gather_main_fields_cyl_linear`.
* Add `rho`, `a` and `aphi` to openpmd output.
* Set correct field position in `z` in openpmd output (node centered
and not cell centered).